Updated creating application load balancer in AWS
Added picture before starting Exercise 5. This will make sure nobody else stumbles because if this setting is on then it's different and not up to date with instructions.
Changed last Exercise #6 (Stop the EC2 Instance) into #8 because step #7 (Stop and Terminate the Resources) was before that.
Fixed typo on Exercise #8 (Stop the EC2 Instance) step 7. first word was "repair" but should say "repeat".